+++ VendaDeAutomoveis/Repository/ConnectionContext/Interfaces/IVendasRepository.cs
@@ -1,8 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 using VendaDeAutomoveis.Entidades;
 
 namespace VendaDeAutomoveis.Repository.ConnectionContext.Interfaces
@@ -11,6 +8,6 @@ namespace VendaDeAutomoveis.Repository.ConnectionContext.Interfaces
     {
         IList<Venda> BuscarPorCliente(Guid? id);
 
-        decimal GastosPorCliente(Guid id);
+        double GastosPorCliente(Guid id);
     }
 }
